Summary:  * Implemented extensible configuration framework.  * Refactored module structure to avoid cyclic dependencies that arose    in the process.
Module Refactoring:
 * Proposed structure:
   * Core modules (not to import out-of-class modules):
       Prelude < Types < Lenses < out-of-class
     Currently this is violated by FocusList and Gtk, which have
     escaped classification and refactor because they haven't
     caused a problem.
   * UI/Exterior modules (not to be imported by out-of-class modules):
         out-of-class < Config; Config.Colour
 * Action taken:
   * Moved data declarations from Config to Types.
     The module now imports and explicitly re-exports them.
   * Cordoned (most) lenses into their own new Lenses module.
   * Moved select functions from Types to their usage point in App.

Extensible Configuration:
 * The basic pieces are written in Types:
   * New classes:
     * ConfigurationExtension
     * TmMessage
   * New data types:
     * ConfigHooks
       * Only has createTermHook field for now; should be extended.
     * SomeConfigExtension
 * Further interface provided by the new module Config.Extension:
   * Re-exports core extension classes and types.
   * Provides the `Both` extension combinator, and the `<+>` operator
     that utilises it to cleanly incorporate an extension into
     existing config.
   * Provides the `sendMessage` and `fromMessage` functions.
 * Config.Colour module refactoring colour configuration as an
   extension.
