+++ src/Termonad/Prelude.hs
@@ -2,7 +2,11 @@
 
 module Termonad.Prelude
   ( module X
+  , whenJust
   ) where
 
 import ClassyPrelude as X
 import Data.Proxy as X
+
+whenJust :: Monoid m => Maybe a -> (a -> m) -> m
+whenJust = flip foldMap
